backend/requirements.txt

Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.0
Pillow==10.1.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
openai==1.3.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
gunicorn==21.2.0